+++ healtryback/src/main/java/com/oscar/healtry/repository/NutricionistaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.oscar.healtry.model.Nutricionista;
 
 @Repository
-public interface NutricionistaRepository extends ListCrudRepository<Nutricionista, Integer> {
+public interface NutricionistaRepository extends ListCrudRepository<Nutricionista, Long> {
 }
